LayoutTests/inspector/debugger/async-stack-trace-basic-expected.txt

Tests for checking that async stack traces exist when pausing in asynchronous callbacks: requestAnimationFrame, setTimeout, setInterval, and postMessage.


== Running test suite: AsyncStackTrace.Basic
-- Running test case: AsyncStackTrace.Basic.Microtask
PAUSED
CALL STACK:
0: [F] handleMicrotask
ASYNC CALL STACK:
1: --- queueMicrotask ---
2: [F] triggerMicrotask
3: [P] Global Code

-- Running test case: AsyncStackTrace.Basic.NestedMicrotask
PAUSED
CALL STACK:
0: [F] handleMicrotask2
ASYNC CALL STACK:
1: --- queueMicrotask ---
2: [F] handleMicrotask1
3: --- queueMicrotask ---
4: [F] triggerNestedMicrotask
5: [P] Global Code

-- Running test case: AsyncStackTrace.Basic.RequestAnimationFrame
PAUSED
CALL STACK:
0: [F] handleAnimationFrame
ASYNC CALL STACK:
1: --- requestAnimationFrame ---
2: [F] triggerRequestAnimationFrame
3: [P] Global Code

-- Running test case: AsyncStackTrace.Basic.NestedRequestAnimationFrame
PAUSED
CALL STACK:
0: [F] handleAnimationFrame2
ASYNC CALL STACK:
1: --- requestAnimationFrame ---
2: [F] handleAnimationFrame1
3: --- requestAnimationFrame ---
4: [F] triggerNestedRequestAnimationFrame
5: [P] Global Code

-- Running test case: AsyncStackTrace.Basic.SetTimeout
PAUSED
CALL STACK:
0: [F] handleTimeout
ASYNC CALL STACK:
1: --- setTimeout ---
2: [F] triggerSetTimeout
3: [P] Global Code

-- Running test case: AsyncStackTrace.Basic.NestedSetTimeout
PAUSED
CALL STACK:
0: [F] handleTimeout2
ASYNC CALL STACK:
1: --- setTimeout ---
2: [F] handleTimeout1
3: --- setTimeout ---
4: [F] triggerNestedSetTimeout
5: [P] Global Code

-- Running test case: AsyncStackTrace.Basic.SetInterval
PAUSED
CALL STACK:
0: [F] handleInterval
ASYNC CALL STACK:
1: --- setInterval ---
2: [F] triggerSetInterval
3: [P] Global Code

-- Running test case: AsyncStackTrace.Basic.NestedSetInterval
PAUSED
CALL STACK:
0: [F] handleInterval2
ASYNC CALL STACK:
1: --- setInterval ---
2: [F] handleInterval1
3: --- setInterval ---
4: [F] triggerNestedSetInterval
5: [P] Global Code

-- Running test case: AsyncStackTrace.Basic.NestedSetInterval
PAUSED
CALL STACK:
0: [F] handleInterval2
ASYNC CALL STACK:
1: --- setInterval ---
2: [F] handleInterval1
3: --- setInterval ---
4: [F] triggerNestedSetInterval
5: [P] Global Code

-- Running test case: AsyncStackTrace.Basic.PostMessage
PAUSED
CALL STACK:
0: [F] handleMessage
ASYNC CALL STACK:
1: --- addEventListener ---
2: [F] triggerPostMessage
3: [P] Global Code

-- Running test case: AsyncStackTrace.Basic.Combined
PAUSED
CALL STACK:
0: [F] handleMessage
ASYNC CALL STACK:
1: --- addEventListener ---
2: [F] triggerPostMessage
3: [F]
4: [F] handleInterval
5: --- setInterval ---
6: [F] triggerSetInterval
7: [F]
8: [F] handleTimeout
9: --- setTimeout ---
10: [F] triggerSetTimeout
11: [F]
12: [F] handleAnimationFrame
13: --- requestAnimationFrame ---
14: [F] triggerRequestAnimationFrame
15: [F]
16: [F] handleMicrotask
17: --- queueMicrotask ---
18: [F] triggerMicrotask
19: [F] triggerCombined
20: [P] Global Code
